requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
groq==0.4.2
gunicorn==22.0.0

flask-login
werkzeug
edge-tts